baseline_model/requirements.txt
ffmpeg-python
fairseq
faiss-cpu
scipy
librosa
tensorboardX
matplotlib